add dedicated folder for color input
